+++ ext/Cache/Cell.svelte
@@ -1,28 +1,40 @@
 <script>
   import { Link } from 'svelte-routing';
+  import Checkbox from '../Checkbox.svelte';
+
   export let row;
   export let col;
+  export let onSelect;
 
+  const shortenKeys = ['host', 'path'];
   const shorten = (str) => {
     if (str.length > 30) {
-      return str.slice(0, 15) + ' ... ' + str.slice(-10);
+      return str.slice(0, 5) + ' ... ' + str.slice(-20);
     }
 
     return str;
   };
   let value, cyID;
-  $: value =
-    col.key === 'host' || col.key === 'path'
-      ? shorten(row[col.key])
-      : row[col.key];
+  $: {
+    const raw = row[col.key] || row.id;
+    value = shortenKeys.includes(col.key) ? shorten(raw) : raw;
+  }
   $: cyID = col.key === 'edit' ? `edit-${row.id}` : col.key;
 </script>
 
 <div class="Cell">
   {#if col.key === 'edit'}
     <Link to="/cache/entry/{row.id}" data-cy-id="cache-cell-{cyID}"
-      >{row.id}</Link
+      >{`${value.slice(0, 15)}...`}</Link
     >
+  {:else if col.key === 'select'}
+    <Checkbox
+      name="select-row-{row.id}"
+      id="select-row-{row.id}"
+      label=""
+      checked={row.checked}
+      onClick={(e) => onSelect(row.id, e.target.checked)}
+    />
   {:else}
     <span data-cy-id="cache-cell-{cyID}">{value}</span>
   {/if}