+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.25 - Cache select all, Remove SideNav, Add global search
+
+- feat: cache select all/multi-select & delete
+- fix: Waterfall highlight
+- remove: SideNav
+- add: global search
+- fix: display pathname without query in cache table
+
 ## 0.0.24 - Fix broken cache file reads
 
 - fix: catch unhandled json parse during reset